+++ examples/apps/todomvc/src/store/index.js
@@ -1,0 +1,11 @@
+import { createStore } from "@inglorious/store"
+import { connectDevTools } from "@inglorious/store/client/dev-tools"
+
+import { entities } from "./entities"
+import { types } from "./types"
+
+export const store = createStore({ types, entities })
+
+if (import.meta.env.DEV) {
+  connectDevTools(store)
+}