+++ packages/renderer-2d/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/renderer-2d
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/engine@4.0.0
+
 ## 2.0.1
 
 ### Patch Changes